requirements.txt
python-mpv
pywebview
qtpy
PyQt6
PyQt6-WebEngine